Added hitboxes to mesh_loading.
Hitboxes are now specified via indices to vertices given by the mesh.
The vertices for a convex set, just the points, no actual faces are
needed, as in a convex set they just are the convex hull, and do not
have to be specified explicitly.

Added hitboxes for the cross.mesh, which was loaded via file.