Relative position in MinkowskiGenerator.
The Generator now generates Minkowski-Vertices with one of the source
objects being offset by a vec3. Since vertices in the provided Sequences
are relative to their own Origin this was necessary. The position is
added to all vertices coming from the second object.

Ive also added the support function for gjk, which is not tested, yet.
It should give the maximum minkowski point in a given direction.