The beginnings of the Input refactoring.
Chapter0:
First stage:
There is not so much change in functionality compared to before.
We map sfml-Events to our own set of InputEvents.
I have added a namespace, called InputEventSpec, where some variables
are set to be drawn upon from related components. These variables try to
implement the specification of InputEventIds, basically providing an ID
for every Input we currently support. If the ID-Generation must happen
according to some rules, the variables set by those rules are to be
located there.
We can therefore also do some test on the adherence of our
implementation to said specification, which I did.
I will however not determine currently, if this test is complete or not.

We now have a component, called a InputEventSource, which encapsulate a
producer of InputEvents for our program. I've decided to implement an
input-iterator at its base, so all the stl-sequence goodies could be
used with input events.
I implemented in a special type for SFML-Events, of course, which
remains to be tested properly.