Added a glmutils function to extract angles from a vector.
Our rotation representation currently is a 3D vector holding a rotation
instruction by axis in radian angles.
If we have a vector, that is a direction, that means the vector points
in a direction which shall be the direction of facing, we need to
convert this vector to its angular instructions from the default facing
direction of +z or (0,0,1).