Serialized the Application Startup.
The Application Startup was (and still is) quite convoluted.
The was an unnecessary thread created in the unnecessary class of
Graphics, which actually did spawn off the whole simulation and did
graphics only as a part of that.... sigh*
Seems to be a relic of old project code.
Nontheless the changes in this branch encompass the deletion of the
class and the relocation of the "functionality" into the main thread
in a way the status quo isn't altered.
From here on the next task is cleanup and sensible abstaction.