Added new time utilities.
They are not integrated.

* Float Seconds
A std::chrono::duration for seconds and subseconds with internal representation of a float.
Used for quick explicit transformation to a float-datatype.

* RelativeClock
A clock that runs in a relative rate to another clock.
The rate is adjustable.
The clock has certain properties and parameters to let it run as a
SimClock

* SFMLClock
A Clock Type that is a wrapper for sfmls Clock utility

* SessionClock
A clock that wraps another clock and is itself aware of system time.
Used for conversions to or from system time.

* SimClock
A specified type of RelativeClock used in Simulations as the time base.

* WallClock
A specified session clock that uphold accuracy requirements for our
purposes.

All of those are easily accesible via the time.hpp header.