Disjointed the SFMLHandle from Application setup.
The SFMLHandle was historically designed to encapsulate SFML-Library
capabilites. However it was horribly misused as the application
launching point in the past.
The SFMLHandle handled the creation of IGameModes/Simulations. This
feature was extracted. In the process, the last global variable
passthrough was exchanged with a sensible parameter passthrough
as it should've been from the start.