ModelObjects showed wrong behavior.
ModelObjects are lazy-computing objects that apply a
matrix transformation to the vertices of a model lazily.
Since the draw-calls to these ModelObjects did not inherently require
transformation of the vertices, but just a pass through of the matrix to
an opengl call, the vertices where never actually updated, which caused
objects to refuse to rotate. This was not only wrong for he pass on to
for example collision routines, but the minimal AABBs used where
actually not computed the way they should've been.

The fix consists of forcing the vertex-update more often. The lazyness
of the component ist thereby basically useless.