Fix for issue #33.
When writing the Operator I encountered inconsistency with the creation
or activation of the graphics context. Where it was instantiated it was
set inactive. This was done to make it available to another thread to
take over drawing. This is however as of now not a thing and is more
inconsistent as the created context is expected to be active. We needed
to activate it then in MainApp.cpp manually, which is weird.
Removed 2 lines that caused the issue.