Started refactoring EventMapper to eradicate c-style.
The whole mapping-events-to-behavior-thing is implemented and used via
lists of indices, that index switch-cased behavior.
The goal is to make an object oriented design out of it, that is easier
usable and expandable.
For that a new type of mapping that uses higher order functions has been
written. Its usage already integrated in parallel but, so far, unused.