Moved keypress triggers into actions.
Actions that concern keypresses have previously gotten a condition
whether to be triggered on keypress or release.
This, however, obstructed us from defining a sensible default
(key_pressed) for example for the toggle action.
This has been reworked, so a controlvalue can now be given to the action
itself and a additional condition is not needed in case of key_pressed.