Moved test out of quarantine and fixed it somewhat.
The somewhat part refers to std::filesystem being inconsistently
implemented across compilers.
Seemingly, SFML depends on MinGw 7.3.0 on Windows, although this IMO needs further
research, which is currently worked on.
This MinGw implementation however implements only
experimental/filesystem and not filesystem ... also does so
inconsistently, so I cannot pass permissions needed on Windows to a
created directory to be able to remove it afterwards.
The accute solution is to simply not delete the directory...
A better compiler, that works on this mess of a platform is currently
searched for.