Implemented trigger via typical eventmapping.
Trigger, which operated Zombie_Gun, was previously implemented as a custom call throughput through all structures from eventhandling to entity. Unacceptable for one-way-event-mapping for the remote control feature.
The feature has been replaced with a simple bool in the control status. While the feature beforehand was super tick accurate, the trigger feature now is only tick accurate and therefore possibly lag prone. This is an explicit downgrade, but one we want to do at this point.
All legacy code regarding the previous trigger feature should be removed.