Separation of Concerns for User.
The User was thought to be the abstraction of a physical User in front of a machine.
We needed to extract some concerns of it into interfaces, so different types of users can be implemented.
This is needed to implement users for synchronized applications, where a User for the Simulation on a Server must be present, the physical User however is through a network connection.

During refactoring this, some other changes were made.
EventHandler was reduced to nothing. #44
It's remains can be found in InputEvent, which describes our EventType and provides a Filter for uniqification of InputEvents. The EventHandler, as a component, was removed.
The Implementation of Window (SFMLWindow) now is a Source for InputEvents, instead of sf::Events.
Realizing this required the window to learn to lock the window, so differential locked mouse movement could be reimplemented.