Fixed up the client considerably.
There were and still are, a lot of issues on the client.
One was that it wouldnt display anything.
Investigations led first to an entity being spawned by the client in it's simulation, which is forbidden.
Surrounding code had to be updated, so it is possible to initialize everything correctly at the right time.
Secondly, by receiving the Avatar/EntityPlayer for the client to control via SyncableManager, we accidentally sliced the EntityPlayer object, which was gruesome.