Improved writing of Header to UdpPacket.
The UdpPacket has a Header that contains timestamps.
The Time of writing these timestamps however is more flexible than the library would provide. We thereby hack around in a packets data buffer.
I improved the already existing hack by using another packet and copying header contents into the buffer of the actual packet to use the packets capability to handle endianess.
I wrote a test_syncprotocol unittest that passes, testing equality of the headers and payload put in and out. And where the respecting memory of the Packets buffer is located.