Gui Main Menu and Perspective Part
There is defined now a Main Menu, where initial flow control of the
application can be inserted. It for now only can switch on the display
of the Imgui Demo.

The Perspective class is the first to get its own Gui element, a decorless window in the top left that displays the Simulation time.